feat: turned conversation into a component
